chore(lint): clear workspace clippy warnings for v1.9.1
`cargo clippy --all-targets -- -D warnings` failed with 57 pre-existing
warnings. Fix them so the release tree passes the local gate:

- Mechanical (clippy --fix + manual): unused imports, derivable Default
  impls, or_default, redundant closures/patterns, matches!, values(),
  collapsed else-if, from_ref, next_back, is_err, flatten nested format!,
  drop a useless `as i64` cast.
- Remove genuinely dead items: gateway engine.rs (unused streaming scaffold
  plus AppState.engine and its main.rs wiring), not_implemented,
  control_error_with_request_id, idempotency_conflict,
  openai_sse_content_delta (now cfg(test)-only), list_disk_alerts alias,
  DeployReject alias, STARTUP_GRACE_MS, Engine::validate_config.
- Mark intentional-but-unused contracts with targeted #[allow] + reason:
  PowerLLM compat request fields, PreparedUpstream.model, Responses stream,
  AppState log_path/bff_url, vLLM max_model_len, too_many_arguments on
  node/scheduler runtime entrypoints, result_large_err on
  evaluate_compatibility.

No behavior change: edits are lint-level or removal of unreachable code.
`./scripts/ci.sh` stays green.

+++ crates/nebula-bff/src/powerllm_compat.rs
@@ -123,7 +123,7 @@ pub async fn token_compat(
         Json(PowerLLMTokenResp {
             access_token: token,
             token_type: "bearer".to_string(),
-            expire_in_minutes: (st.session_ttl_hours as i64) * 60,
+            expire_in_minutes: st.session_ttl_hours * 60,
         }),
     ))
 }
@@ -145,6 +145,7 @@ pub async fn user_info_compat() -> impl IntoResponse {
 // ---------------------------------------------------------------------------
 
 #[derive(Debug, Deserialize)]
+#[allow(dead_code)] // query params accepted for PowerLLM client compatibility; not all are filtered on yet
 pub struct InstanceListQuery {
     pub model_name: Option<String>,
     pub model_uid: Option<String>,
@@ -346,6 +347,7 @@ pub async fn get_instance_detail_compat(
 // ---------------------------------------------------------------------------
 
 #[derive(Debug, Deserialize)]
+#[allow(dead_code)] // accepted for PowerLLM client compatibility; unused fields are ignored
 pub struct LaunchInstanceCompatReq {
     pub model_uid: String,
     pub model_name: Option<String>,
